Update LiveDataExt.kt
Combination of a `?.let {}` with `?: run` can lead to unexpected behaviour. Since `let` returns its result, if the last expression of `let` resolves to null, then `run` is also executed. So you end up with both parts executed. `Apply` on the other hand returns `this` value, so no matter what let resolves to, run is only executed if the initial value is null